feat(FOROME-1310): fix comments - create components from util functions
